pyproject.toml
[project]
name = "fast-next"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "google-generativeai>=0.8.5",
    "openai-agents>=0.0.14",
    "pyperclip>=1.9.0",
    "python-dotenv>=1.1.0",
    "uvicorn>=0.34.2",
]
